refactor: DTO 변환 레이어 변경
