+++ src/main/java/chess/controller/GameController.java
@@ -1,12 +1,15 @@
 package chess.controller;
 
 import chess.dto.*;
+import chess.utils.Serializer;
 import chess.service.GameService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @Controller
 public class GameController {
@@ -25,26 +28,31 @@ public class GameController {
     @GetMapping("/chessboard/{gameId}")
     @ResponseBody
     public ChessBoardDTO loadGame(@PathVariable String gameId) {
-        return gameService.loadGame(gameId);
+        return Serializer.deserializeAsDTO(gameService.loadGame(gameId));
     }
 
     //    gameid를 int로 받을지 string으로 받을지
     @GetMapping("/turn/{gameId}")
     @ResponseBody
     public TurnDTO turn(@PathVariable String gameId) {
-        return gameService.turn(gameId);
+        return new TurnDTO(gameService.turn(gameId));
     }
 
     //여기서 DTO요리
     @PutMapping(path = "/move/{gameId}")
     public ResponseEntity move(@PathVariable String gameId, @RequestBody MoveDTO moveDTO) {
-        return gameService.move(gameId, moveDTO);
+        try {
+            gameService.move(gameId, moveDTO);
+            return new ResponseEntity(HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity(HttpStatus.BAD_REQUEST);
+        }
     }
 
     @GetMapping("/finishById/{gameId}")
     @ResponseBody
     public FinishDTO isFinished(@PathVariable String gameId) {
-        return gameService.isFinished(gameId);
+        return new FinishDTO(gameService.isFinished(gameId));
     }
 
     @PostMapping("/finish/{gameId}")
@@ -56,12 +64,13 @@ public class GameController {
     @GetMapping("/scoreById/{gameId}")
     @ResponseBody
     public ScoreDTO score(@PathVariable String gameId) {
-        return gameService.score(gameId);
+        List<Double> scores = gameService.score(gameId);
+        return new ScoreDTO(scores.get(0), scores.get(1));
     }
 
     @PostMapping("/restart/{gameId}")
     @ResponseBody
     public ChessBoardDTO restart(@PathVariable String gameId) {
-        return gameService.restart(gameId);
+        return Serializer.deserializeAsDTO(gameService.restart(gameId));
     }
 }
