+++ src/main/java/chess/service/GameService.java
@@ -3,12 +3,15 @@ package chess.service;
 import chess.domain.board.Position;
 import chess.domain.game.ChessGame;
 import chess.domain.piece.Color;
-import chess.dto.*;
+import chess.dto.MoveDTO;
+import chess.dto.ScoreDTO;
 import chess.repository.ChessRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 
 @Service
 public class GameService {
@@ -19,27 +22,21 @@ public class GameService {
         this.chessRepository = chessRepository;
     }
 
-    public ChessBoardDTO loadGame(String gameId) {
-        return chessRepository.loadGameAsDTO(gameId);
+    public String loadGame(String gameId) {
+        return chessRepository.loadGame(gameId);
     }
 
-    public TurnDTO turn(String gameId) {
+    public String turn(String gameId) {
         return chessRepository.turn(gameId);
     }
 
-    //controller 에 트라이캐치
-    public ResponseEntity move(String gameId, MoveDTO moveDTO) {
-        try {
-            ChessGame chessGame = chessRepository.loadGameById(gameId);
-            Position sourcePosition = Position.of(moveDTO.getSource());
-            Position targetPosition = Position.of(moveDTO.getTarget());
-            chessGame.move(sourcePosition, targetPosition);
-            checkGameOver(gameId, chessGame);
-            chessRepository.saveGame(gameId, chessGame);
-            return new ResponseEntity(HttpStatus.OK);
-        } catch (Exception e) {
-            return new ResponseEntity(HttpStatus.BAD_REQUEST);
-        }
+    public void move(String gameId, MoveDTO moveDTO) {
+        ChessGame chessGame = chessRepository.loadGameById(gameId);
+        Position sourcePosition = Position.of(moveDTO.getSource());
+        Position targetPosition = Position.of(moveDTO.getTarget());
+        chessGame.move(sourcePosition, targetPosition);
+        checkGameOver(gameId, chessGame);
+        chessRepository.saveGame(gameId, chessGame);
     }
 
     private void checkGameOver(String gameId, ChessGame chessGame) {
@@ -48,7 +45,7 @@ public class GameService {
         }
     }
 
-    public FinishDTO isFinished(String gameId) {
+    public boolean isFinished(String gameId) {
         return chessRepository.isFinishedById(gameId);
     }
 
@@ -56,14 +53,14 @@ public class GameService {
         chessRepository.finish(gameId);
     }
 
-    public ScoreDTO score(String gameId) {
+    public List<Double> score(String gameId) {
         ChessGame chessGame = chessRepository.loadGameById(gameId);
-        return new ScoreDTO(chessGame.getScore(Color.BLACK), chessGame.getScore(Color.WHITE));
+        return new ArrayList(Arrays.asList(chessGame.getScore(Color.BLACK), chessGame.getScore(Color.WHITE)));
     }
 
-    public ChessBoardDTO restart(String gameId) {
+    public String restart(String gameId) {
         ChessGame chessGame = new ChessGame();
         chessRepository.restart(gameId, chessGame);
-        return chessRepository.loadGameAsDTO(gameId);
+        return chessRepository.loadGame(gameId);
     }
 }
