+++ src/main/java/chess/service/LobbyService.java
@@ -32,10 +32,7 @@ public class LobbyService {
 
     public Boolean isDuplicate(String title) {
         Optional<String> foundRoom = chessRepository.findRoomId(title);
-        if (foundRoom.isPresent()) {
-            return true;
-        }
-        return false;
+        return foundRoom.isPresent();
     }
 
     public ChessGame loadGame(String roomName) {
